pyproject.toml
[project]
name = "chlamydia_16s_dais"
version = "0.1.0"
description = "NASBA primer validation framework for Chlamydia trachomatis 16S rRNA detection"
authors = [
    {name = "gil", email = "gil.harari@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.0,<4.0.0"

[tool.poetry]
name = "chlamydia_16s_dais"
version = "0.1.0"
description = "NASBA primer validation framework for Chlamydia trachomatis 16S rRNA detection"
authors = ["gil <gil.harari@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
numpy = ">=1.21.0,<3.0.0"
biopython = "^1.80"
click = "^8.0"
tqdm = "^4.60"
nupack = { path = "./dependencies/nupack-4.0.1.12-cp313-cp313-linux_x86_64.whl" }
#nupack = { path = "./dependencies/nupack-4.0.2.0-cp313-cp313-linux_x86_64.whl" }
poetry-core = ">=1.0.0"
pandas-stubs = "==2.3.0.250703"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
